Extract the filter bar into its own view
