+++ Pinwheel/Sources/Pinwheel/Catalog/PinwheelCatalogView.swift
@@ -190,14 +190,6 @@ private struct PinwheelIndexView: SwiftUI.View {
     @State private var selectedTag: PinTag?
     @State private var scrolledDistance: CGFloat = 0
 
-    // Height matches the offset so the scrim sits flush below the pill bar.
-    private static let scrimHeight: CGFloat = 40
-    private static let scrimRampDistance: CGFloat = 24
-
-    private var fadeOpacity: Double {
-        Double(min(1, max(0, scrolledDistance) / Self.scrimRampDistance))
-    }
-
     var body: some SwiftUI.View {
         ScrollViewReader { proxy in
             ZStack(alignment: .trailing) {
@@ -262,56 +254,14 @@ private struct PinwheelIndexView: SwiftUI.View {
         }
         .safeAreaInset(edge: .top, spacing: 0) {
             if sectionTags.count > 1 {
-                filterBar
-                    .overlay(alignment: .bottom) {
-                        LinearGradient(
-                            colors: [
-                                PinwheelTheme.Colors.primaryBackground,
-                                PinwheelTheme.Colors.primaryBackground.opacity(0)
-                            ],
-                            startPoint: .top,
-                            endPoint: .bottom
-                        )
-                        .frame(height: Self.scrimHeight)
-                        .offset(y: Self.scrimHeight)
-                        .opacity(fadeOpacity)
-                        .allowsHitTesting(false)
-                    }
+                PinwheelFilterBar(
+                    tags: sectionTags,
+                    selectedTag: $selectedTag,
+                    scrolledDistance: scrolledDistance
+                )
             }
         }
         .onChange(of: section?.id) { selectedTag = nil }
-    }
-
-    private var filterBar: some SwiftUI.View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: .spacingS) {
-                ForEach(sectionTags, id: \.self) { tag in
-                    pill(title: tag.rawValue, isSelected: selectedTag == tag) {
-                        selectedTag = selectedTag == tag ? nil : tag
-                    }
-                }
-            }
-            .padding(.horizontal, .spacingS)
-            .padding(.vertical, .spacingS)
-        }
-        .background(.primaryBackground)
-    }
-
-    private func pill(title: String, isSelected: Bool, action: @escaping () -> Void) -> some SwiftUI.View {
-        SwiftUI.Button(action: action) {
-            PinLabel(title)
-                .color(isSelected ? .custom(PinwheelTheme.Colors.primaryBackground) : .primary)
-                .padding(.horizontal, .spacingM)
-                .padding(.vertical, .spacingS)
-                .background {
-                    if isSelected {
-                        Capsule().fill(.actionText)
-                    } else {
-                        Capsule().strokeBorder(.secondaryText, lineWidth: 1)
-                    }
-                }
-        }
-        .buttonStyle(.plain)
     }
 
     private var sectionTags: [PinTag] {
